[LS-001-B] Add test for deletion of user [BJ, PT]
