+++ app/src/androidTest/java/com/example/loginsesame/DatabaseTest.kt
@@ -1,0 +1,51 @@
+package com.example.loginsesame
+
+import android.content.Context
+import androidx.room.Room
+import androidx.test.core.app.ApplicationProvider
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.example.loginsesame.data.User
+import com.example.loginsesame.data.UserDao
+import com.example.loginsesame.data.UserDatabase
+import org.junit.After
+import org.junit.Assert
+import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+import java.io.IOException
+import kotlin.jvm.Throws
+
+@RunWith(AndroidJUnit4::class)
+class SimpleEntityReadWriteTest {
+    private lateinit var userDao: UserDao
+    private lateinit var db: UserDatabase
+
+    @Before
+    fun createDb() {
+        val context = ApplicationProvider.getApplicationContext<Context>()
+        db = Room.inMemoryDatabaseBuilder(
+                context, UserDatabase::class.java).build()
+        userDao = db.getUserDao()
+    }
+
+    @After
+    @Throws(IOException::class)
+    fun closeDb() {
+        db.close()
+    }
+
+    @Test
+    @Throws(Exception::class)
+    fun writeUserAndReadInList() {
+        val user = User(1, "Max Musterman", "test@mail.com", "123456789")
+        val rowid = userDao.insertUser(user)
+        Assert.assertNotNull(rowid)
+        val createdUsers = userDao.getAllUsers()
+        assert(createdUsers.size == 1)
+
+        userDao.deleteUser(user)
+        val deletedUsers = userDao.getAllUsers()
+        assert(deletedUsers.isEmpty())
+    }
+
+}
