+++ app/src/androidTest/java/com/example/loginsesame/ExampleInstrumentedTest.kt
@@ -35,32 +35,3 @@ class ExampleInstrumentedTest {
 
 
 }
-
-@RunWith(AndroidJUnit4::class)
-class SimpleEntityReadWriteTest {
-    private lateinit var userDao: UserDao
-    private lateinit var db: UserDatabase
-
-    @Before
-    fun createDb() {
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        db = Room.inMemoryDatabaseBuilder(
-            context, UserDatabase::class.java).build()
-        userDao = db.getUserDao()
-    }
-
-    @After
-    @Throws(IOException::class)
-    fun closeDb() {
-        db.close()
-    }
-
-    @Test
-    @Throws(Exception::class)
-    fun writeUserAndReadInList() {
-        val user: User = User(1, "Max Musterman", "test@mail.com", "123456789");
-        val rowid = userDao.insertUser(user)
-        assertNotNull(rowid)
-    }
-
-}
